update: moved login, register out from auth context and removed navigation
